double changed to int for number if bins
git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@6832 0381ead4-6506-0410-b988-94b70fbc4730

+++ littrack/base/CbmLitFieldFitter.h
@@ -50,8 +50,8 @@ public:
 
 	void SetXangle(double xangle) {fXangle = xangle;}
 	void SetYangle(double yangle) {fYangle = yangle;}
-	void SetNofBinsX(double nofBinsX) {fNofBinsX = nofBinsX;}
-	void SetNofBinsY(double nofBinsY) {fNofBinsY = nofBinsY;}
+	void SetNofBinsX(int nofBinsX) {fNofBinsX = nofBinsX;}
+	void SetNofBinsY(int nofBinsY) {fNofBinsY = nofBinsY;}
 	void SetUseEllipseAcc(bool useEllipseAcc) {fUseEllipseAcc = useEllipseAcc;}
 
 private:
